Refactor voucher feature: Remove old voucher screen, add discount and my voucher pages, and update asset images
